chore(node): move out fxutil into libs
